refactor: State 를 독립된 클래스로 분리
