+++ src/main/java/com/project/undefined/job/entity/Stage.java
@@ -37,10 +37,6 @@ public class Stage extends BaseEntity {
     @Enumerated(value = EnumType.STRING)
     private State state;
 
-    public enum State {
-        NONE, PASS, FAIL, WAIT
-    }
-
     public static Stage of(final String name, final Job job) {
         return new Stage(null, name, job, State.NONE);
     }
