+++ src/main/java/com/project/undefined/job/entity/State.java
@@ -1,0 +1,9 @@
+package com.project.undefined.job.entity;
+
+public enum State {
+    NONE, PASS, FAIL, WAIT;
+
+    public static State from(final String str) {
+        return State.valueOf(str.toUpperCase());
+    }
+}
